ver listado de articulos. Crear, editar y eliminar.

+++ CarritoDeCompras/src/main/java/models/Articulo.java
@@ -1,11 +1,12 @@
 package models;
 
+
 public class Articulo {
 	int id;
-	double codigo;
+	int codigo;
 	String nombre; 
 	String descripcion;
-	double precio;
+	Double precio;
 	int stock;
 
 	
@@ -13,7 +14,8 @@ public class Articulo {
 		super();
 	}
 
-	public Articulo(double codigo, String nombre, String descripcion, double precio, int stock) {
+
+	public Articulo(int codigo, String nombre, String descripcion, double precio, int stock) {
 		super();
 		this.codigo = codigo;
 		this.nombre = nombre;
@@ -23,55 +25,76 @@ public class Articulo {
 	}
 
 
-
-
-
 	public int getId() {
 		return id;
 	}
+
 
 	public void setId(int id) {
 		this.id = id;
 	}
 
-	public double getCodigo() {
+
+	public int getCodigo() {
 		return codigo;
 	}
-	public void setCodigo(double codigo) {
+
+
+	public void setCodigo(int codigo) {
 		this.codigo = codigo;
 	}
+
+
 	public String getNombre() {
 		return nombre;
 	}
+
+
 	public void setNombre(String nombre) {
 		this.nombre = nombre;
 	}
+
+
 	public String getDescripcion() {
 		return descripcion;
 	}
+
+
 	public void setDescripcion(String descripcion) {
 		this.descripcion = descripcion;
 	}
-	public double getPrecio() {
+
+
+	public Double getPrecio() {
 		return precio;
 	}
-	public void setPrecio(double precio) {
+
+
+	public void setPrecio(Double precio) {
 		this.precio = precio;
 	}
+
+
 	public int getStock() {
 		return stock;
 	}
+
+
 	public void setStock(int stock) {
 		this.stock = stock;
 	}
 
 
-	
 	@Override
 	public String toString() {
-		return "Articulo [id=" + id + ", codigo=" + codigo + ", nombre=" + nombre + ", descripcion=" + descripcion + ", precio="
-				+ precio + ", stock=" + stock + "]";
+		return "Articulo [id=" + id + ", codigo=" + codigo + ", nombre=" + nombre + ", descripcion=" + descripcion
+				+ ", precio=" + precio + ", stock=" + stock + "]";
 	}
 	
-
+	
+	
+	
+	
+	
+	
 }
